pyproject.toml
[project]
name = "my-finance-api"
version = "0.1.0"
description = "Backend API for financial data analysis and ML model serving"
readme = "README.md"
requires-python = "~=3.11.0"
dependencies = [
    "fastapi>=0.128.0",
    "google-cloud-bigquery>=3.39.0",
    "google-cloud-storage>=3.7.0",
    "httpx>=0.28.1",
    "ipykernel>=7.1.0",
    "pandas-gbq>=0.33.0",
    "polymodel==2.1.0",
    "pre-commit>=4.5.1",
    "python-jose>=3.5.0",
    "requests>=2.31.0",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "duckdb>=1.0.0",
    "keyring>=25.7.0",
    "keyrings-google-artifactregistry-auth>=1.1.2", # Does not work for god knows why, thus using tokens
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[[tool.uv.index]]
name = "private"
url = "https://europe-north1-python.pkg.dev/rasmus-prod/python-packages/simple/"

[tool.uv.sources]
polymodel = { index = "private" }

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["logs*", "model_artifacts*", "tests*", "scripts*"]

[tool.bandit]
# Exclude B105: hardcoded password string - for OAuth2 standard strings and error codes
# Exclude B106: hardcoded password function argument - for OAuth2 "Bearer" token type
skips = ["B105", "B106"]
